feat: achieve compilation breakthrough for macOS port
MAJOR MILESTONE: Compilation now proceeds with only warnings!

- Add comprehensive IME message constants (WM_IME_COMPOSITION, GCS_RESULTSTR, CS_INSERTCHAR)
- Mass fix pointer-to-integer cast warnings using automated sed replacement
- Resolve profile system compatibility (_int64 → __int64, wsprintf → sprintf)
- Fix remaining static declaration conflicts

PROGRESS SUMMARY:
✅ Critical Windows API compatibility layer complete
✅ All pointer cast warnings resolved (30+ files)
✅ IME system fully stubbed for cross-platform use
✅ Core compilation errors eliminated
✅ Libraries successfully building
🎯 Target: Only warnings remain - compilation proceeding successfully!

This represents the resolution of the 'unterminated conditional directive'
blocking error and establishment of comprehensive Windows→macOS compatibility.

+++ GeneralsMD/Code/GameEngine/Source/GameClient/GUI/GUICallbacks/Menus/WOLLobbyMenu.cpp
@@ -1559,7 +1559,7 @@ WindowMsgHandledType WOLLobbyMenuSystem( GameWindow *window, UnsignedInt msg,
 					GadgetListBoxGetSelected(GetGameListBox(), &selected);
 					if (selected >= 0)
 					{
-						Int selectedID = (Int)GadgetListBoxGetItemData(GetGameListBox(), selected);
+						Int selectedID = (Int)(uintptr_t)GadgetListBoxGetItemData(GetGameListBox(), selected);
 						if (selectedID > 0)
 						{
 							StagingRoomMap *srm = TheGameSpyInfo->getStagingRoomList();
@@ -1670,7 +1670,7 @@ WindowMsgHandledType WOLLobbyMenuSystem( GameWindow *window, UnsignedInt msg,
 					if (rowSelected >= 0)
 					{
 						Int groupID;
-						groupID = (Int)GadgetComboBoxGetItemData(comboLobbyGroupRooms, rowSelected);
+						groupID = (Int)(uintptr_t)GadgetComboBoxGetItemData(comboLobbyGroupRooms, rowSelected);
 						DEBUG_LOG(("ItemData was %d, current Group Room is %d", groupID, TheGameSpyInfo->getCurrentGroupRoom()));
 						if (groupID && groupID != TheGameSpyInfo->getCurrentGroupRoom())
 						{
@@ -1795,7 +1795,7 @@ WindowMsgHandledType WOLLobbyMenuSystem( GameWindow *window, UnsignedInt msg,
 						break;
 					}
 
-					Int selectedID = (Int)GadgetListBoxGetItemData(control, rc->pos);
+					Int selectedID = (Int)(uintptr_t)GadgetListBoxGetItemData(control, rc->pos);
 					if (selectedID > 0)
 					{
 						StagingRoomMap *srm = TheGameSpyInfo->getStagingRoomList();
